Move and add to /user/dashboard, add some display CSS
